+++ src/configurations/RGBTConnectConfig.cpp
@@ -6,5 +6,5 @@
 
 unsigned long RGBTConnectConfig::MAX_NUM_ITER   = 1e9;
 unsigned long RGBTConnectConfig::MAX_NUM_STATES = 10000;
-float RGBTConnectConfig::MAX_PLANNING_TIME      = 60e3;
+int RGBTConnectConfig::MAX_PLANNING_TIME        = 60e3;
 int RGBTConnectConfig::NUM_LAYERS               = 5;
